Adding delivery manager

+++ VendrediProto/Assets/Component/Island/Scripts/ScriptableObjects/IslandSO.cs
@@ -7,23 +7,23 @@ public class IslandSO : ScriptableObject
 {
     [SerializeField] private string _islandName;
     [SerializeField] private FactionTypes _factionType;
-    [SerializeField] private SerializableDictionary<MerchandiseSO, int> _merchandisesToSell;
-    [SerializeField] private SerializableDictionary<MerchandiseSO, int> _merchandisesRequested;
-    [SerializeField] private float _merchandiseRequestedTimeInterval = 220f;
-    [SerializeField] private float _merchandiseToTakeTimeInterval = 180f;
+    [SerializeField] private SerializableDictionary<MerchandiseSO, ushort> _merchandisesToSell;
+    [SerializeField] private SerializableDictionary<MerchandiseSO, ushort> _merchandisesRequested;
+    [SerializeField] private uint _merchandiseRequestedTimeInterval = 220;
+    [SerializeField] private uint _merchandiseToTakeTimeInterval = 180;
 
     public string IslandName => _islandName;
     public FactionTypes FactionType => _factionType;
-    public SerializableDictionary<MerchandiseSO, int> MerchandisesToSell => _merchandisesToSell;
-    public SerializableDictionary<MerchandiseSO, int> MerchandisesRequested => _merchandisesRequested;
+    public SerializableDictionary<MerchandiseSO, ushort> MerchandisesToSell => _merchandisesToSell;
+    public SerializableDictionary<MerchandiseSO, ushort> MerchandisesRequested => _merchandisesRequested;
     public float MerchandiseRequestedTimeInterval => _merchandiseRequestedTimeInterval;
     public float MerchandiseToSellTimeInterval => _merchandiseToTakeTimeInterval;
 
-    public (MerchandiseType, int, int) RequestRandomMerchandiseRequest()
+    public (MerchandiseType, ushort, uint) RequestRandomMerchandiseRequest()
     {
         int randomIndex = Random.Range(0, _merchandisesRequested.ToDictionary().Count);
         var randomRequest = _merchandisesRequested.ToDictionary().ElementAt(randomIndex);
 
-        return (randomRequest.Key.Type, randomRequest.Value, (int)_merchandiseRequestedTimeInterval);
+        return (randomRequest.Key.Type, randomRequest.Value, _merchandiseRequestedTimeInterval);
     }
 }
